+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Tideland Go Library
 
+## 2015-06-25
+
+- Done migration into new library
+- Added new configuration package
+
 ## 2015-06-05
 
 - Started migration of existing packages into new library